chore: initialization touchups
