+++ pkg/cloudproviders/aws/apis/v1alpha1/tags.go
@@ -22,8 +22,9 @@ import (
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/samber/lo"
 
-	"github.com/aws/karpenter-core/pkg/apis/provisioning/v1alpha5"
 	"github.com/aws/karpenter/pkg/utils/injection"
+
+	"github.com/aws/karpenter-core/pkg/apis/provisioning/v1alpha5"
 )
 
 func MergeTags(ctx context.Context, custom ...map[string]string) (result []*ec2.Tag) {
