+++ pkg/cloudproviders/aws/cloudprovider/instancetype.go
@@ -35,7 +35,7 @@ import (
 	"github.com/aws/karpenter/pkg/cloudproviders/aws/apis/v1alpha1"
 	"github.com/aws/karpenter/pkg/cloudproviders/aws/cloudprovider/amifamily"
 	"github.com/aws/karpenter/pkg/cloudproviders/common/cloudprovider"
-	"github.com/aws/karpenter/pkg/utils/injection"
+	"github.com/aws/karpenter/pkg/operator"
 	"github.com/aws/karpenter/pkg/utils/resources"
 )
 
@@ -69,8 +69,8 @@ func NewInstanceType(ctx context.Context, info *ec2.InstanceTypeInfo, kc *v1alph
 	instanceType.maxPods = instanceType.computeMaxPods(ctx, kc)
 
 	// Precompute to minimize memory/compute overhead
-	instanceType.resources = instanceType.computeResources(injection.GetOptions(ctx).AWSEnablePodENI)
-	instanceType.overhead = instanceType.computeOverhead(injection.GetOptions(ctx).VMMemoryOverhead, kc)
+	instanceType.resources = instanceType.computeResources(operator.GetOptions(ctx).AWSEnablePodENI)
+	instanceType.overhead = instanceType.computeOverhead(operator.GetOptions(ctx).VMMemoryOverhead, kc)
 	instanceType.requirements = instanceType.computeRequirements()
 	return instanceType
 }
@@ -370,7 +370,7 @@ func (i *InstanceType) computeMaxPods(ctx context.Context, kc *v1alpha5.KubeletC
 	switch {
 	case kc != nil && kc.MaxPods != nil:
 		mp = ptr.Int64(int64(ptr.Int32Value(kc.MaxPods)))
-	case !injection.GetOptions(ctx).AWSENILimitedPodDensity:
+	case !operator.GetOptions(ctx).AWSENILimitedPodDensity:
 		mp = ptr.Int64(110)
 	default:
 		mp = ptr.Int64(i.eniLimitedPods())
